requirements.txt
statsmodels
tqdm
scipy
seaborn
pandas
numpy
matplotlib
biopython
